Refactor test setup code
