--- packages/automerge-repo-react-hooks/test/testSetup.ts
@@ -1,7 +1,0 @@
-import "@testing-library/jest-dom"
-import { cleanup } from "@testing-library/react"
-import { afterEach } from "vitest"
-
-afterEach(() => {
-  cleanup()
-})
